Feat: cloudinary dependency and settings
It was installed cloudinary dependency to set the configuration of cloudinary, for this was added new variables in the file .env and .env example. Then was implemented this variables in the file configEnv.ts, after that in the file app.ts was implemented the cloudinary feature. Finally was created 2 new files in the folder shared/globals/helpers/cloudinary ; cloudinaryUploads.ts and fileResult.interface.ts