Feat: new component,  features and dependencies
It were installed new dependencies wich were; multer and mongoose. It was create a new component called archiveRoutes.ts which was implemented inside the component route.ts. It were edit and added new feature in the file archive.ts. It was added new feature in the component JoiValidation.decorators.ts which features were implemented in the controller archive.ts. It was added a feature in the component cloudinaryUploads.ts for can able to upload files to cloudinary. It was created a new generator in the file generators.ts. It was fix a bug in the file archive.service.ts which was implemented in the controller archive.ts. It was modified some features and added new ones to the controller archive.ts. Finally was created a new env variable and was added to the file configEnv.ts